+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Nodi family: properly restore the "Low CPU" option when it is enabled in either module in really old patches.
 
+- Vimina: timer manipulation stops working after about 10 minutes.
+
 
 ---